Added CalibratedGpioPinAnalogInputSafe and SafeAdcInputMeasurementFailure interface. Implemented both in the main procedure. Compiles, but NOT TESTED. Not even loaded to the RPI.
The theory behind this implementation was because of the failure mode where the ADC or pressure transducer goes bad. By the old, simplified logic the pump would turn on and never turn off, either burning out the pump or over pressurizing the system, or both. The new "safe" routine extends the old routine and checks for a rapid change in the pressure (configurable at runtime). If the change in pressure is experienced then a special callback is used to handle the error. Right now, the callback shuts down the gpio, prints some error logging and then spins forever. 

In the future it needs to indicate visually that there is an error, notify the user, run a diagnostic, and, if safe, return itself to operation.