Removed the halt on rapid pressure change and reduced the max pressure from 62 to 58.
This was done due to excessive overpressure detection and repeated
and continuing system shutdown.

There's something else going on causing this but I don't know what it
is yet and we need water.

Also, now the MQTT client is going nuts and reconnecting constantly for
a reason I can't figure out yet.