Reinstated the MQTT client with fewer calls to mqttClient.Connect().
It seems to have fixed the problem but we'll need to watch out for it.

Also changed the pressure error routine to check for overpressure
instead of delta-pressure. Remember, if an open circuit failure mode
occurs on the pressure sensor we get something like 91 PSI. We should
never see much more than 70 psi (which is already high and we need to 
figure out why we're swinging 8-12 PSI every time we turn the pump on
or off).